Consolidated Drug object
Drug object, while necessary, did not fully implement everything that was required of it if it were to be introduced into a repository (e.g. did not display stock). Therefore, I created DrugStock; a class that contains Drug as an internal class. Additionally, due to the number of combinations of constructors of the object, I felt it was necessary to introduce the Builder pattern.