Implemented I_UniqueQueryableRepository and I_Unique
By introducing I_UniqueQueryableRepository and I_Unique, repository controllers can now be queried based on a unique the stored objects have. Additionally, this meant that I could create IDFactory; a factory responsible for the creation logic on IDs. Furthermore, the way I've created the interfaces provides some flexibility for future implementations of unique variables because of the use of generics in the interfaces declarations.