Prevented Drug duplicates
Added I_UniqueQueryableRepository and I_Unique to DrugRepositoryController and DrugStock respectively. This is to ensure that no drugs with exactly the same names are entered into the system. If a drug with the same name is added to the system, it must have a different description. If it fails to meet this criteria, a DuplicateDrugException is thrown.