Altered I_SerialisationStrategy
I altered the serialisation strategy interface because it didn't make sense for you to be able to serialise something with one strategy but deserialise it with another. I_SerialisationStrategy and I_DeserialisationStrategy, and DefaultSerialisationStrategy and DefaultDeserialisationStrategy have been consolidated to I_SerialisationStrategy and DefaultSerialisationStrategy respectively. These changes have been reflected in RepositorySerialisationController.